+++ domain/test_account.py
@@ -1,5 +1,5 @@
-from domain.account import (
-    Account,
+from domain.bankaccount import (
+    BankAccount,
     InsufficientFundsException
 )
 from domain.transaction import create_transaction
@@ -7,7 +7,7 @@ from domain.transaction import create_transaction
 
 def test_transfer1():
     t = create_transaction(1, 2, 20)
-    x = Account(2, [t])
+    x = BankAccount(2, [t])
     y = 3
     x.transfer(y, 10)
     assert x.get_balance() == 10
@@ -15,7 +15,7 @@ def test_transfer1():
 
 def test_transfer_poor_ryan():
     t = create_transaction(1, 2, 10)
-    x = Account(2, [t])
+    x = BankAccount(2, [t])
     y = 3
     try:
         x.transfer(y, 20)
@@ -26,13 +26,13 @@ def test_transfer_poor_ryan():
 
 def test_transactions_static_size():
     t = create_transaction(1, 2, 10)
-    x = Account(2, [t])
+    x = BankAccount(2, [t])
     assert len(x._transactions) == 1
 
 
 def test_transactions_incremented_size():
     t = create_transaction(1, 2, 10)
-    x = Account(2, [t])
+    x = BankAccount(2, [t])
     y = 3
     x.transfer(y, 10)
     assert len(x._transactions) == 2
@@ -40,7 +40,7 @@ def test_transactions_incremented_size():
 
 def test_no_balance():
     t = create_transaction(1, 2, 10)
-    x = Account(2, [t])
+    x = BankAccount(2, [t])
     y = 3
     x.transfer(y, 10)
     assert x.get_balance() == 0
